+++ dkg-primitives/Cargo.toml
@@ -20,6 +20,7 @@ serde = { version = "1.0", features = ["derive"] }
 bincode = "1.3.3"
 rand = "0.8.4"
 chacha20poly1305 = "0.9.0"
+typed-builder = { git = "https://github.com/idanarye/rust-typed-builder.git", branch = "master" }
 
 curv = { package = "curv-kzen", version = "0.9", default-features = false }
